+++ src/BotApi.php
@@ -3,8 +3,12 @@
 namespace TelegramBot\Api;
 
 use TelegramBot\Api\Http\CurlHttpClient;
+use TelegramBot\Api\Types\UserChatBoosts;
+use TelegramBot\Api\Types\ReplyParameters;
 use TelegramBot\Api\Http\HttpClientInterface;
 use TelegramBot\Api\Types\ArrayOfBotCommand;
+use TelegramBot\Api\Types\LinkPreviewOptions;
+use TelegramBot\Api\Types\ArrayOfReactionType;
 use TelegramBot\Api\Types\ArrayOfChatMemberEntity;
 use TelegramBot\Api\Types\ArrayOfMessageEntity;
 use TelegramBot\Api\Types\ArrayOfMessages;
@@ -337,6 +341,8 @@ class BotApi
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
+     * @param LinkPreviewOptions|null $linkPreviewOptions Link preview generation options for the message.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -352,19 +358,42 @@ class BotApi
         $disableNotification = false,
         $messageThreadId = null,
         $protectContent = null,
-        $allowSendingWithoutReply = null
+        $allowSendingWithoutReply = null,
+        $replyParameters = null,
+        $linkPreviewOptions = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
+        if (null === $linkPreviewOptions && false !== $disablePreview) {
+            @trigger_error('setting $disablePreview is now deprecated use $linkPreviewOptions instead', E_USER_DEPRECATED);
+
+            $linkPreviewOptions = new LinkPreviewOptions();
+            $linkPreviewOptions->map([
+                'is_disabled' => $disablePreview
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendMessage', [
             'chat_id' => $chatId,
             'text' => $text,
             'message_thread_id' => $messageThreadId,
             'parse_mode' => $parseMode,
-            'disable_web_page_preview' => $disablePreview,
-            'reply_to_message_id' => (int) $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson(),
+            'link_preview_options' => is_null($linkPreviewOptions) ? $linkPreviewOptions : $linkPreviewOptions->toJson()
         ]));
     }
 
@@ -381,6 +410,7 @@ class BotApi
      * @param InlineKeyboardMarkup|ReplyKeyboardMarkup|ReplyKeyboardRemove|ForceReply|null $replyMarkup
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return MessageId
      * @throws Exception
@@ -399,8 +429,22 @@ class BotApi
         $allowSendingWithoutReply = false,
         $replyMarkup = null,
         $messageThreadId = null,
-        $protectContent = null
+        $protectContent = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return MessageId::fromResponse($this->call('copyMessage', [
             'chat_id' => $chatId,
             'from_chat_id' => $fromChatId,
@@ -410,10 +454,9 @@ class BotApi
             'caption_entities' => $captionEntities,
             'disable_notification' => (bool) $disableNotification,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => (int) $replyToMessageId,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'protect_content' => (bool) $protectContent,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -430,6 +473,7 @@ class BotApi
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws Exception
@@ -444,19 +488,32 @@ class BotApi
         $disableNotification = false,
         $messageThreadId = null,
         $protectContent = null,
-        $allowSendingWithoutReply = null
+        $allowSendingWithoutReply = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendContact', [
             'chat_id' => $chatId,
             'phone_number' => $phoneNumber,
             'first_name' => $firstName,
             'last_name' => $lastName,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -636,6 +693,7 @@ class BotApi
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      *
@@ -651,19 +709,32 @@ class BotApi
         $livePeriod = null,
         $messageThreadId = null,
         $protectContent = null,
-        $allowSendingWithoutReply = null
+        $allowSendingWithoutReply = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendLocation', [
             'chat_id' => $chatId,
             'latitude' => $latitude,
             'longitude' => $longitude,
             'live_period' => $livePeriod,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -753,6 +824,7 @@ class BotApi
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws Exception
@@ -769,8 +841,22 @@ class BotApi
         $disableNotification = false,
         $messageThreadId = null,
         $protectContent = null,
-        $allowSendingWithoutReply = null
+        $allowSendingWithoutReply = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendVenue', [
             'chat_id' => $chatId,
             'latitude' => $latitude,
@@ -779,11 +865,10 @@ class BotApi
             'address' => $address,
             'foursquare_id' => $foursquareId,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -798,6 +883,7 @@ class BotApi
      * @param bool $protectContent Protects the contents of the sent message from forwarding and saving
      * @param bool $allowSendingWithoutReply Pass True if the message should be sent even if the specified replied-to message is not found
      * @param string|null $messageThreadId
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -811,17 +897,30 @@ class BotApi
         $disableNotification = false,
         $protectContent = false,
         $allowSendingWithoutReply = false,
-        $messageThreadId = null
+        $messageThreadId = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendSticker', [
             'chat_id' => $chatId,
             'sticker' => $sticker,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -1060,6 +1159,7 @@ class BotApi
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
      * @param \CURLFile|\CURLStringFile|string|null $thumbnail
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -1078,22 +1178,35 @@ class BotApi
         $messageThreadId = null,
         $protectContent = null,
         $allowSendingWithoutReply = null,
-        $thumbnail = null
+        $thumbnail = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendVideo', [
             'chat_id' => $chatId,
             'video' => $video,
             'duration' => $duration,
             'caption' => $caption,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'supports_streaming' => (bool) $supportsStreaming,
             'parse_mode' => $parseMode,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
             'thumbnail' => $thumbnail,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -1114,6 +1227,7 @@ class BotApi
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
      * @param \CURLFile|\CURLStringFile|string|null $thumbnail
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -1131,21 +1245,34 @@ class BotApi
         $messageThreadId = null,
         $protectContent = null,
         $allowSendingWithoutReply = null,
-        $thumbnail = null
+        $thumbnail = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendAnimation', [
             'chat_id' => $chatId,
             'animation' => $animation,
             'duration' => $duration,
             'caption' => $caption,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'parse_mode' => $parseMode,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
             'thumbnail' => $thumbnail,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -1169,6 +1296,7 @@ class BotApi
      * @param string|null $parseMode
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -1185,20 +1313,33 @@ class BotApi
         $allowSendingWithoutReply = false,
         $parseMode = null,
         $messageThreadId = null,
-        $protectContent = null
+        $protectContent = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendVoice', [
             'chat_id' => $chatId,
             'voice' => $voice,
             'caption' => $caption,
             'duration' => $duration,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
-            'allow_sending_without_reply' => $allowSendingWithoutReply,
             'parse_mode' => $parseMode,
             'protect_content' => (bool) $protectContent,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -1260,6 +1401,7 @@ class BotApi
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
      * @param \CURLFile|\CURLStringFile|string|null $thumbnail
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -1281,21 +1423,34 @@ class BotApi
         $parseMode = null,
         $protectContent = null,
         $allowSendingWithoutReply = null,
-        $thumbnail = null
+        $thumbnail = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendAudio', [
             'chat_id' => $chatId,
             'audio' => $audio,
             'duration' => $duration,
             'performer' => $performer,
             'title' => $title,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'parse_mode' => $parseMode,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
             'thumbnail' => $thumbnail,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -1312,6 +1467,7 @@ class BotApi
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -1327,19 +1483,32 @@ class BotApi
         $parseMode = null,
         $messageThreadId = null,
         $protectContent = null,
-        $allowSendingWithoutReply = null
+        $allowSendingWithoutReply = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendPhoto', [
             'chat_id' => $chatId,
             'photo' => $photo,
             'caption' => $caption,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'parse_mode' => $parseMode,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -1358,6 +1527,7 @@ class BotApi
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
      * @param \CURLFile|\CURLStringFile|string|null $thumbnail
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -1374,20 +1544,33 @@ class BotApi
         $messageThreadId = null,
         $protectContent = null,
         $allowSendingWithoutReply = null,
-        $thumbnail = null
+        $thumbnail = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendDocument', [
             'chat_id' => $chatId,
             'document' => $document,
             'caption' => $caption,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'parse_mode' => $parseMode,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
             'thumbnail' => $thumbnail,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -1612,6 +1795,7 @@ class BotApi
      * @param string|null $parseMode
      * @param bool $disablePreview
      * @param InlineKeyboardMarkup|null $replyMarkup
+     * @param LinkPreviewOptions|null $linkPreviewOptions Link preview generation options for the message.
      *
      * @return Message|true
      * @throws Exception
@@ -1623,8 +1807,18 @@ class BotApi
         $parseMode = null,
         $disablePreview = false,
         $replyMarkup = null,
-        $inlineMessageId = null
+        $inlineMessageId = null,
+        $linkPreviewOptions = null
     ) {
+        if (null === $linkPreviewOptions && false !== $disablePreview) {
+            @trigger_error('setting $disablePreview is now deprecated use $linkPreviewOptions instead', E_USER_DEPRECATED);
+
+            $linkPreviewOptions = new LinkPreviewOptions();
+            $linkPreviewOptions->map([
+                'is_disabled' => $disablePreview
+            ]);
+        }
+
         $response = $this->call('editMessageText', [
             'chat_id' => $chatId,
             'message_id' => $messageId,
@@ -1804,6 +1998,7 @@ class BotApi
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws Exception
@@ -1834,8 +2029,22 @@ class BotApi
         $sendEmailToProvider = false,
         $messageThreadId = null,
         $protectContent = null,
-        $allowSendingWithoutReply = null
+        $allowSendingWithoutReply = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendInvoice', [
             'chat_id' => $chatId,
             'title' => $title,
@@ -1855,14 +2064,13 @@ class BotApi
             'need_email' => $needEmail,
             'need_shipping_address' => $needShippingAddress,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'provider_data' => $providerData,
             'send_phone_number_to_provider' => (bool) $sendPhoneNumberToProvider,
             'send_email_to_provider' => (bool) $sendEmailToProvider,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -2371,6 +2579,7 @@ class BotApi
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
      * @param \CURLFile|\CURLStringFile|string|null $thumbnail
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws InvalidArgumentException
@@ -2387,20 +2596,33 @@ class BotApi
         $messageThreadId = null,
         $protectContent = null,
         $allowSendingWithoutReply = null,
-        $thumbnail = null
+        $thumbnail = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendVideoNote', [
             'chat_id' => $chatId,
             'video_note' => $videoNote,
             'duration' => $duration,
             'length' => $length,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => $replyToMessageId,
             'reply_markup' => is_null($replyMarkup) ? $replyMarkup : $replyMarkup->toJson(),
             'disable_notification' => (bool) $disableNotification,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
             'thumbnail' => $thumbnail,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -2416,6 +2638,7 @@ class BotApi
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
      * @param array<string, \CURLFile|\CURLStringFile> $attachments Attachments to use in attach://<attachment>
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message[]
      * @throws Exception
@@ -2428,16 +2651,29 @@ class BotApi
         $messageThreadId = null,
         $protectContent = null,
         $allowSendingWithoutReply = null,
-        $attachments = []
+        $attachments = [],
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return ArrayOfMessages::fromResponse($this->call('sendMediaGroup', [
             'chat_id' => $chatId,
             'media' => $media->toJson(),
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => (int) $replyToMessageId,
             'disable_notification' => (bool) $disableNotification,
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ] + $attachments));
     }
 
@@ -2463,6 +2699,7 @@ class BotApi
      * @param int|null $messageThreadId Unique identifier for the target message thread (topic) of the forum; for forum supergroups only
      * @param bool|null $protectContent
      * @param bool|null $allowSendingWithoutReply
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws Exception
@@ -2483,8 +2720,22 @@ class BotApi
         $replyMarkup = null,
         $messageThreadId = null,
         $protectContent = null,
-        $allowSendingWithoutReply = null
+        $allowSendingWithoutReply = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendPoll', [
             'chat_id' => $chatId,
             'question' => $question,
@@ -2496,10 +2747,9 @@ class BotApi
             'is_closed' => (bool) $isClosed,
             'disable_notification' => (bool) $disableNotification,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => (int) $replyToMessageId,
             'reply_markup' => $replyMarkup === null ? $replyMarkup : $replyMarkup->toJson(),
             'protect_content' => (bool) $protectContent,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -2522,6 +2772,7 @@ class BotApi
      *                                                                                                  keyboard or to force a reply from the user.
      * @param int|null $messageThreadId
      * @param bool|null $protectContent
+     * @param ReplyParameters|null $replyParameters Description of the message to reply to.
      *
      * @return Message
      * @throws Exception
@@ -2536,17 +2787,30 @@ class BotApi
         $allowSendingWithoutReply = false,
         $replyMarkup = null,
         $messageThreadId = null,
-        $protectContent = null
+        $protectContent = null,
+        $replyParameters = null
     ) {
+        if (null !== $replyToMessageId || null !== $allowSendingWithoutReply) {
+            @trigger_error(
+                'setting $replyToMessageId or $allowSendingWithoutReply is now deprecated use $replyParameters instead',
+                E_USER_DEPRECATED
+            );
+
+            $replyParameters = new ReplyParameters();
+            $replyParameters->map([
+                'message_id' => $replyToMessageId,
+                'allow_sending_without_reply' => (bool) $allowSendingWithoutReply
+            ]);
+        }
+
         return Message::fromResponse($this->call('sendDice', [
             'chat_id' => $chatId,
             'emoji' => $emoji,
             'disable_notification' => (bool) $disableNotification,
             'message_thread_id' => $messageThreadId,
-            'reply_to_message_id' => (int) $replyToMessageId,
-            'allow_sending_without_reply' => (bool) $allowSendingWithoutReply,
             'reply_markup' => $replyMarkup === null ? $replyMarkup : $replyMarkup->toJson(),
             'protect_content' => (bool) $protectContent,
+            'reply_parameters' => is_null($replyParameters) ? $replyParameters : $replyParameters->toJson()
         ]));
     }
 
@@ -2769,6 +3033,152 @@ class BotApi
         }
 
         return $this;
+    }
+
+    /**
+     * Use this method to change the chosen reactions on a message.
+     * Service messages can't be reacted to.
+     * Automatically forwarded messages from a channel to its discussion group have the same available reactions as messages in the channel.
+     *
+     * @param string|int $chatId Unique identifier for the target chat or username of the target channel (in the format @channelusername)
+     * @param int $messageId Identifier of the target message.
+     * @param ArrayOfReactionType $reaction A list of reaction types to set on the message.
+     * @param bool $isBig Pass `true` to set the reaction with a big animation
+     *
+     * @return bool
+     * @throws Exception
+     *
+     * @author bernard-ng <bernard@devscast.tech>
+     */
+    public function setMessageReaction($chatId, $messageId, $reaction, $isBig = false)
+    {
+        return $this->call('setMessageReaction', [
+            'chat_id' => $chatId,
+            'message_id' => $messageId,
+            'reaction' => $reaction,
+            'is_big' => $isBig
+        ]);
+    }
+
+    /**
+     * Use this method to delete multiple messages simultaneously.
+     * If some of the specified messages can't be found, they are skipped.
+     * Returns True on success.
+     *
+     * @param string|int $chatId Unique identifier for the target chat or username of the target channel (in the format @channelusername)
+     * @param int[] $messageIds A JSON-serialized list of 1-100 identifiers of messages to delete. See deleteMessage for limitations on which messages can be deleted
+     *
+     * @return bool
+     * @throws Exception
+     *
+     * @author bernard-ng <bernard@devscast.tech>
+     */
+    public function deleteMessages($chatId, $messageIds)
+    {
+        return $this->call('deleteMessages', [
+            'chat_id' => $chatId,
+            'message_ids' => $messageIds
+        ]);
+    }
+
+    /**
+     * Use this method to forward multiple messages of any kind.
+     * If some of the specified messages can't be found or forwarded, they are skipped.
+     * Service messages and messages with protected content can't be forwarded.
+     * Album grouping is kept for forwarded messages.
+     * On success, an array of MessageId of the sent messages is returned.
+     *
+     * @param string|int $chatId Unique identifier for the target chat or username of the target channel (in the format @channelusername)
+     * @param string|int $fromChatId Unique identifier for the chat where the original messages were sent (or channel username in the format @channelusername)
+     * @param int[] $messageIds A JSON-serialized list of 1-100 identifiers of messages in the chat from_chat_id to forward. The identifiers must be specified in a strictly increasing order.
+     * @param bool $disableNotification Sends the messages silently. Users will receive a notification with no sound.
+     * @param int|null $messageThreadId Unique identifier for the target message thread (topic) of the forum; for forum supergroups only
+     * @param bool $protectContent Protects the contents of the forwarded messages from forwarding and saving
+     *
+     * @return int[]
+     * @throws Exception
+     *
+     * @author bernard-ng <bernard@devscast.tech>
+     */
+    public function forwardMessages(
+        $chatId,
+        $fromChatId,
+        $messageIds,
+        $messageThreadId = null,
+        $disableNotification = false,
+        $protectContent = false
+    ) {
+        return $this->call('forwardMessages', [
+            'chat_id' => $chatId,
+            'from_chat_id' => $fromChatId,
+            'message_ids' => $messageIds,
+            'message_thread_id' => $messageThreadId,
+            'disable_notification' => (bool) $disableNotification,
+            'protect_content' => (bool) $protectContent
+        ]);
+    }
+
+    /**
+     * Use this method to copy messages of any kind.
+     * If some of the specified messages can't be found or copied, they are skipped.
+     * Service messages, paid media messages, giveaway messages, giveaway winners messages, and invoice messages can't be copied
+     *
+     * A quiz poll can be copied only if the value of the field correct_option_id is known to the bot.
+     * The method is analogous to the method forwardMessages, but the copied messages don't have a link to the original message. Album grouping is kept for copied messages.
+     *
+     * On success, an array of MessageId of the sent messages is returned.
+     *
+     * @param string|int $chatId Unique identifier for the target chat or username of the target channel (in the format @channelusername)
+     * @param string|int $fromChatId Unique identifier for the chat where the original messages were sent (or channel username in the format @channelusername)
+     * @param int[] $messageIds A JSON-serialized list of 1-100 identifiers of messages in the chat from_chat_id to copy. The identifiers must be specified in a strictly increasing order.
+     * @param int|null $messageThreadId Unique identifier for the target message thread (topic) of the forum; for forum supergroups only
+     * @param bool $disableNotification Sends the messages silently. Users will receive a notification with no sound.
+     * @param bool $protectContent Protects the contents of the copied messages from forwarding and saving
+     * @param bool $removeCaption Pass True to copy the messages without their captions
+     *
+     * @return int[]
+     * @throws Exception
+     *
+     * @author bernard-ng <bernard@devscast.tech>
+     */
+    public function copyMessages(
+        $chatId,
+        $fromChatId,
+        $messageIds,
+        $messageThreadId,
+        $disableNotification = false,
+        $protectContent = false,
+        $removeCaption = false
+    ) {
+        return $this->call('copyMessages', [
+            'chat_id' => $chatId,
+            'from_chat_id' => $fromChatId,
+            'message_ids' => $messageIds,
+            'message_thread_id' => $messageThreadId,
+            'disable_notification' => (bool) $disableNotification,
+            'protect_content' => (bool) $protectContent,
+            'remove_caption' => (bool) $removeCaption
+        ]);
+    }
+
+    /**
+     * Use this method to get the list of boosts added to a chat by a user. Requires administrator rights in the chat.
+     * Returns a UserChatBoosts object.
+     *
+     * @param string|int $chatId Unique identifier for the chat or username of the channel (in the format @channelusername)
+     * @param int $userId Unique identifier of the target user
+     *
+     * @return UserChatBoosts
+     * @throws Exception
+     *
+     * @author bernard-ng <bernard@devscast.tech>
+     */
+    public function getUserChatBoosts($chatId, $userId)
+    {
+        return UserChatBoosts::fromResponse($this->call('getUserChatBoosts', [
+            'chat_id' => $chatId,
+            'user_id' => $userId
+        ]));
     }
 
     /**