+++ tests/Types/UpdateTest.php
@@ -39,6 +39,10 @@ class UpdateTest extends AbstractTypeTest
             'poll_answer' => PollAnswerTest::getMinResponse(),
             'poll' => PollTest::getMinResponse(),
             'chat_join_request' => ChatJoinRequestTest::getMinResponse(),
+            'message_reaction' => MessageReactionUpdatedTest::getMinResponse(),
+            'message_reaction_count' => MessageReactionCountUpdatedTest::getMinResponse(),
+            'chat_boost' => ChatBoostUpdatedTest::getMinResponse(),
+            'chat_boost_removed' => ChatBoostRemovedTest::getMinResponse(),
         ];
     }
 
@@ -63,6 +67,10 @@ class UpdateTest extends AbstractTypeTest
         $this->assertNull($item->getMyChatMember());
         $this->assertNull($item->getChatMember());
         $this->assertNull($item->getChatJoinRequest());
+        $this->assertNull($item->getMessageReaction());
+        $this->assertNull($item->getMessageReactionCount());
+        $this->assertNull($item->getChatBoost());
+        $this->assertNull($item->getChatBoostRemoved());
     }
 
     /**
@@ -84,5 +92,9 @@ class UpdateTest extends AbstractTypeTest
         $this->assertEquals(PollAnswerTest::createMinInstance(), $item->getPollAnswer());
         $this->assertEquals(PollTest::createMinInstance(), $item->getPoll());
         $this->assertEquals(ChatJoinRequestTest::createMinInstance(), $item->getChatJoinRequest());
+        $this->assertEquals(MessageReactionUpdatedTest::createMinInstance(), $item->getMessageReaction());
+        $this->assertEquals(MessageReactionCountUpdatedTest::createMinInstance(), $item->getMessageReactionCount());
+        $this->assertEquals(ChatBoostUpdatedTest::createMinInstance(), $item->getChatBoost());
+        $this->assertEquals(ChatBoostRemovedTest::createMinInstance(), $item->getChatBoostRemoved());
     }
 }